Added new vscode extensions and shortcuts. Ninja upload function fixed.

+++ .config/fish/functions/ninja.fish
@@ -1,5 +1,6 @@
 function ninja --description 'Upload a file or directory to tmp.ninja'
     if test (count $argv) -ne 1
+        set_color brred
         echo 'Please provide a file to upload'
         return
     end
@@ -11,15 +12,25 @@ function ninja --description 'Upload a file or directory to tmp.ninja'
     else
         set fileUri (realpath $argv)
     end
+
     echo -e 'Uploading...\n'
-    set -l res (curl -F file=@$fileUri 'https://tmp.ninja/api.php?d=upload-tool')
-    set_color brgreen
-    echo -e "\nUploaded $argv to: $res"
-    echo "Upload URL copied to clipboard"
-    set_color normal
-    if test -e $zipName
-        echo -e '\nRemoving zipped directory...'
-        rm -rf $zipName
+    set -l response (curl -F files[]=@$fileUri 'https://tmp.ninja/upload.php')
+
+    if test (echo $response | jq -r .success) = true
+        set uploadUrl (echo $response | jq .files[0].url)
+        set_color brgreen
+        echo -e "\nUploaded $argv to: $uploadUrl"
+        echo "Upload URL copied to clipboard"
+        set_color normal
+        if test -e $zipName
+            echo -e '\nRemoving zipped directory...'
+            rm -rf $zipName
+        end
+        echo -n $uploadUrl | pbcopy
+    else
+        set_color brred
+        set error (echo $response | jq -r .description)
+        echo -e "\nError: $error"
     end
-    echo -n $res | pbcopy
+
 end
